Reduce the amount of space used by locations when dumping AST.
Fixes issue #280.

Each location node adds 6 lines to the output of `--dump=ast` or `--dump=checker`.
To avaoid this, we format the location when dumping the AST using the `string.format`
function. To be precise, we try to format for all the fields that end with "loc".
For example, something like:

```
    col_loc = {
      col = 10,
      file_name = "foo.pln",
      line = 1,
      pos = 10
    },
    end_loc = {
      col = 10,
      file_name = "foo.pln",
      line = 1,
      pos = 10
    },
```

becomes:

```
	col_loc = "1:10"
	end_loc = "1:10"
```