Added the --show-locations flag.
Locations in the AST dump take quite a bit of space. According t Hugo, an example showed that
~20% of the lines were "loc" keys, given every single AST node carries a location. Although
there are scenarious where showing locations is useful, it clutters the dump mostly. One example
where the locations are useful is if somebody was extending the translator or something that
needs to know the locations. One compromise we settled on was to add an extra flag `--show-locations`.
If someone is working on the Pallene to Lua translator they could use `--show-locations` to see
the locations. But if they are not then locations would be hidden by default.