The translator now removes cast expressions inside initializer lists.
The translator spec was extended to include a unit test for the same.

The AST node for if statement is a linked-list of if nodes, with the last node an instance of
`ast.Stat.Block`. We previously used while loop to traverse the nodes. However, the
`then_` and `else_` fields are statements that are no different from other kinds of statements.
Therefore, now we call `translate_stat` recursively to traverse these constructs.