The translator now visits ast.Var nodes.
At a very basic level what the `translate_*()` functions are doing is traversing the AST tree, in
search of type annotations. Since the AST IR is built from several mutually recursive types (`Exp`,
`Stat`, `Var`, `Decl`, ...) we also need to use several mutually recursive functions to implement
the traversal -- `translate_exp`, `translate_stat`, etc. In this commit, we introduce yet another
mutually recusrive method `translate_var`. It receives an `ast.Var` node as input and then calls
`translate_exp` on the child nodes when needed to trigger the recursion. This method basically
traverses things that can appear on the LHS of an assignment. For example: `x`, `x[i]`, `x.name`,
and so on.