The translator now generates forward declarations for all the functions.
In this commit, the translator is being prepared to solve the mutually recursive function problem.
The solution is we create local variables before the first function in the set of mutually recursive
functions. Each of these local variables are then assigned an anonymous function. To make way for
this solution, the translator has been extended to generate forward declarations for all the functions.
Obviously, at a later stage we will limit this generation to only functions that are mutually recursive.

Two utility methods `handle_mrf_name` and `erase_mrf_modifier` were created to accomodate this.
Here, `mrf` is an abbreviation of mutually recursive function. The arithmetic involved in adjusting
the indexes were figured out after a session of trial and error.