The translator now generates forward declarations for all the functions.
Since shadowing top-level components is a syntax error in Pallene, the translator can generate all
the forward references at the beginning of the module. This design allows us to not worry about
finding empty lines, lines with comments, and so on to correctly translate mutually recursive
function groups.

Whenever the translator encounters a function it tracks the function name in the `functions` sequence.
Thus, at the end of the translation pass, we will have the list of all the functions. Using this
list we generate the forward declarations in the first line. The translator ensures that any code
related to the forward declarations are generate if and only if there is at least one function in
the module.

Before the partials sequence is converted to a string, the translator creates an auxillary sequence
to generate the forward declarations. After which, the auxillary sequence is prepended to the
partials the sequence. This inserts the forward declarations at the beginning of the module.