The translator now generates the initial space characters before the forward declarations.
Adding indentation to the long strings makes the input and output look different when forward declarations
are generated.  It is also a bit difficult to keep track of all the whitespace when writing the tests.
This problem was solved in this commit by modifying the translator to insert the forward declaration after
the initial space characters, if the first line starts with whitespace.