Updated the translator to generate functions more naturally.
If we generate `function f()` instead of `f = function()`, we not need to use `name_loc`.
`function f` without a `local` in front of it is totally equivalent to `f = function()`.
In other words, if `f` is declared to be a local variable, when you write `function f()`
it assigns to that local variable instead of creating a global function. In Lua, the
`function f()` syntax is syntactic sugar for `f = function()`. It is a good idea to use
the `function f()` syntax in the translator because it would let us get rid of the `name_loc`.
Further, even if we did not have to remove `name_loc`, the `function f()` syntax looks more
natural.