Indetation in the test cases were removed in translator spec.
Removing indentation from the cases allows us to get rid of the complicated code in `add_forward_declarations`
It would spare us from having to do all the extra work of messing with the first partial. Another thing is
that there are some test cases that you are using double quotes instead of long strings because the indentation
of the return statement generated by the translator would look weird.

```
        assert_translation(
            "export i : integer = 0\n" ..
            "\n" ..
            "export function f()\n" ..
            "end",
            "local f; local  i           = 0\n" ..
            "\n" ..
            "f = function ()\n" ..
            "end\n" ..
            "return {\n" ..
            "    i = i,\n" ..
            "    f = f,\n" ..
            "}\n")
```

Since we got rid of the extra indentation in the test suite, this allows us to use long strings for these test cases.