Added utility functions in translator spec to run the translated module.
The main premise of Pallene is that it is compatible with Lua. This compatibility is defined
as removing type annotations from a Pallene program results in transforming it into a Lua
program. The Pallene to Lua translator will allow us to check whether this property is still
valid. We need to automate this validation through unit tests that execute the resulting Lua
programs produced by the translator.

The utility functions introduced in this PR will help us design test cases to verify the
resulting Lua program produced by the translator to work as expected.