Moved the execution test cases from coder spec.
Fixes issue 293. In order to check if the Lua code generated by the Lua backend produces the same
output as the C backend, we decided to reuse the tests from the coder spec. This allows us to
check if the code generated by both the backends work the same way.

In Pallene, the `string` module is accessed via the `string_` keyword. Therefore, the translator
now generates an alias for it.

The C backend adds runtime checks. However, the Lua backend does not generate any such checks.
Therefore, the `execution_tests.run` function accepts the backend identifier for which we are
running the tests. Therefore, it skips any tests that are designed to validate runtime checks.