Get rid of the tofloat builtin
Encourage multiplying by `1.0` instead of using a special `tofloat` builtin. Closes #296.

The tofloat builtin is the only builtin operator that existed in Pallene but not on Lua. This meant
that it was one corner case where erasing the type annotations from Pallene did not result in a
valid Lua program. Since this builtin function is not strictly necessary, I think the best thing to
do is to just get rid of it.

The recommended idiom is to multiply by `1.0` because the C compiler can optimize that as a no-op.
Adding `0.0` also should work however GCC is not able to optimize that as a no-op. I think the
reason it doesn't do so is that `x + 0.0 == x` is not valid for negative zero: `-0.0 + 0.0 == 0.0`.
However, the identity should be valid for any floating point number that was originally an integer.
Clang can optimize that, at least.