Created utility assertion functions for conditionally disabling test cases.
The first thing the execution tests do is compile the Pallene code into a `__test__` module.
Depending on the backend we are testing we generate a `__test__.so` or a `__test__.lua`.
After that, when we call `run_test` we generate a `__test_script__.lua` that tests the code,
calling assert a bunch of times.

One pattern that is common is testing for Pallene runtime errors. Previously we did the follwing
by hand:
```
	local ok, err = pcall(target, arguments)
	assert(not ok)
	assert(string.find(err, "message", nil, true))
```

In order to get rid of this repetitive code, the `assert_pallene_error` was introduced in this
commit. Now we are able to get rid of the boilerplate code with the `assert(string.find)`.
Further, we are now clearly marking the test cases that are testing for Pallene runtime errors,
which are not tested for the Lua backend. In other words, the `assert_pallene_code` function
for the Lua backend is implemented a no-op that skips the test.

Another utility function `assert_pallene_type` was introduced to make assertions that involve
Pallene type checks. It should be noted that there are a few cases where both `assert_pallene_error`
and `assert_pallene_type` cannot be used. In such cases, we manually check the backend we are
testing before executing the tests.

Since the translator was updated to generate `local string_ = string;`, the translator spec was
updated to accomodate this change.