Reworked optimization option in pallenec
Now the option is a flag like GCC optimization flag: -O<level>.
It accepts the same levels as a C compiler and only level "0"
disables Pallene compiler optimizations. There is also an option
for setting specific C compiler flags with the CFLAGS environment
variable that overrides the pallenec flag for the C compiler only.