Don't set exps[i] to false in Stat.Decl
In the checker, when a variable declaration did not have an initializer, we would set the
initializer to `false`. While this is more explicit, it is a bit confusing because it is different
from what the parser returns ans also because someone iterating over stat.exps needs to check if the
exp is not false.

The choice is to either document this behavior in checker.lua or to get rid of it. I think it
getting rid of it is the simplest solution.