Updated the Pallene runtime to report error locations consistently.
In this PR, we also checked if there were any errors that were missing the error location. At first
we checked only the functions that end with `_error`. The following functions did not generate an
error location:
```
pallene_runtime_arity_error
pallene_runtime_record_nonstr_error
pallene_runtime_record_index_error
```

It turned out that, the `pallene_runtime_record_nonstr_error` and `pallene_runtime_record_index_error`
were unused. Apparently, they were used in a previous version of Pallene and are not required anymore.
Therefore, they were deleted.

> Also, if you are working on the .c files, always remember to run make! Otherwise, your changes will
> not be reflected in the runtime.

As for `pallene_runtime_arity_error`, the problem is that we can show the location but we do not have
a way to show the location from where it is being invoked. We have left `pallene_runtime_arity_error`
as is for now.