Added the --tokens flag.
The `--tokens` flag can be used to dump the tokens recognized by the lexer. The compiler exists
immediately after dumping the tokens.

Further, the lexer was updated to include the end location of tokens in this commit. This goes
to say that we do not really have to save comments to satisfy the requirements of the translator.