Updated the translator to erase the regions marked by the parser.
The parser marks all the regions that can be erased by the translator. Basically, a sequence of
ranges is returned from the parser embedded in the AST. This required a few modifications to the
following phases and the tests. The new architecture allows the translator to erase type annotations
without traversing nodes in a recursive fashion.

Previously, we manually took care of:
 * Replacing "export" with "local" for variables
 * Leave "local" as is for variables
 * Removed both "export" and "local" for functions because of the forward declarations

However, now we do not have such flexiblity to manually take care of this. We could remove
export for variables. As for functions, remove local and export altogether. All of this could be
done by just marking the boundaries of the modifier token in the parser. As for `local` in a variable
declaration, we would have to pop a region off, so `local` for local variables is not erased.
This would have looked a bit awkward.

Here is the code segment that marks the modifier.

```
if self:peek("local") or self:peek("export") then
    self:region_begin()
    visibility = self:e()
    self:region_end()
else
    visibility = false
end
```

Since we mark the modifier, it will be erased in the translator. So another alternative was to
not mark the modifier for deletion. We could handle it separately in the translator. The node location
would point to the start of the modifier anyway. But we would have to traverse the AST for that.
But we are already traversing the list of toplevel nodes to prepare the list of forward declarations.
The motivation here is to erase the modifiers in the translator, close to where we
are preparing the forward declarations, instead of splitting the work between the translator and the
parser. To implement this, we would have to find the modifier token boundaries and add them to the
sequence of regions we intend to erase. The problem is, the sequence of regions have to be sorted
(that's how we get the regions from the parser). Why is this necessary? Because the partials are
generated from top to bottom. So we would have to sort the sequence of regions.

However, we came up with another alternative which took the best of both the previous options.
We generate forward declarations for all the symbols regardless of them being exported or not.
Thus, the parser could mark the modifier for removal for both variables and functions. This commit
implements this technique.