Put the comment_regions back in the parser
If we are going to show comments in the --dump=lexer output, then we might as well always return
comments instead of having a configuration switch to sometimes return and sometimes skip comments.
And if we are doing that then we no longer have a good reason to keep the comment_regions in the
lexer.