Allow primitive type names to be used as identifiers
And also allow string.char and string.sub, getting rid of the string_ workaround. Fixes #240.

The first part of this patch is that most builtin types are no longer a keyword, meaning that they
can now be used in variable names. The only exception is `nil`, which was already a keyword in Lua.

The next change that we did is that we got rid of the string_ workaround. The solution we chose was
to add the `string` name in the scope as a module name, but still have it count as the String() type
if it is used as a type name. Currently this workaround is just for the string type / module but in
theory we could extend it to other types in the future if there is a desire to have modules that
export a type with the same name as the module name.