Simplify the benchmarks_spec a bit
The function that we were calling assert_benchmark isn't really an assert because it isn't intended
to be called inside an it() block. I renamed it to test_benchmark. At the same time, I also took the
opportunity to move the describe() block inside of test_benchmark. Normally we don't do this kind of
thing because the stack traces for all the test point to the same line number but in this case it is
OK because all the test cases have clear names in the describe() and it().