Fix location line numbers for comments
Fixes #324.

Since LPEG's match method operates in anchored mode, it was only recognizing the newline if it
happened right at the start of the matched token. We first observed the bug on comments but after
debugging it I think it also applied to lines that contain whitespace at the end of the line.