Simplify how reformat() deals with blank lines
Now the logic for blank lines in the toplevel is the same as inside a function. In both cases blank
lines are ignored by default and /**/ is used to signal an intentional blank line.

We also no longer treat tabs specially. The price for this is that the reformat function is no
longer idempotent but I think that is a fair price to pay to simplify things.