Replace Travis CI by Github Actions
Closes #336.

Travis is apparently disabling or throttling the free CI minutes for OSS projects[1][2].
As a response, I'm migrating our CI to Github Actions.

While I was doing this, I also figured out how to create separate checkmarks for the linter and for
the test suite. If one of them fails, you'll be able to know how to see which one it was in the pull
request page.

Furthermore, I also figured out how to tweak the repo's branch protection rules to warn if a branch
is not rebased on top of master before it it merged. That will help with our workflow, which expects
that the current branch is rebased before it is merged.

[1] https://blog.travis-ci.com/oss-announcement
[2] https://news.ycombinator.com/item?id=25338983