Don't discard extra values in the RHS
Currently, extra arguments in the RHS are discarded during type-checking step and are never
evaluated. This doesn't match Lua's behavior. Those extra arguments might have side-effects and
should still be evaluated.

Another related bug was Pallene's behavior when a function returning multiple values was used in a
function call or in a return statement. Pallene generated code that discarded extra return values,
depending on the types of function that is receiving the values from "g". This also does not match
Lua's behavior. Here, we don't have the option to allow the extra arguments so we raise a compile
type type error instead.

    -- Assume that "g" returns multiple values
    f(g())
    return g()

Fixes #347