Refactor the checker.Name
The old checker.Name was tightly coupling the concerns of the type checker with the concerns of the
IR generator. To improve this situation I am splitting it into two types:
- checker.Symbol (type-related info, for type checking)
- checker.Def (location-related annotations, for the IR generator)

I also removed some features to simplify the code:
- Exported variables. Now you can only export functions
- Mutually recursive functions. (These should be re-implemented in a later commit)

I still need to refactor the IR generator. The IR-related tests are currently broken.