Clean up the parser tests
Many of the AST tests were not pulling their own weight and were prone to breaking after unrelated
changes to the compiler. To keep the test suite more manageable, I decided to delete them.

One kind of test I deleted were test cases asserting that the AST has a specified shape. For a large
number of these tests, there are other test cases in the execution tests that effectively test the
same thing. I kept only the tests where examining the AST is specially helpful, like the ones for
operator precedence.

Another kind of test I deleted were the ones for simple errors like missing tokens or unexpected
token. Most of these tests were added in the old implementation, to test the error labels generated
by our LPegLabel parser. However, the new parser no longer uses LPegLabel, rendering most of these
tests obsolete. I kept the more "interesting" tests, which I could give a nice "it" description, and
deleted the rest.

In addition to deleting tests, I also reordered them to help ensure that the test suite is covering
all the cases. The test are now ordered in the same order as the parser.lua and it's easier to see
if there is a missing test case.

I reindented the AST in a style that is more consistent with the rest of our code base, with 4
spaces for indentation. Previously, there were multiple styles being used to indent the tables.

Finally, I also renamed the "describe" and "it" blocks in more idiomatic Busted style.