Get rid of types.T.Void
These types that aren't actual types are a constant source of trouble, because when combined with
type inference there are some situations where they can survive all the way to the coder without
raising an error.

My solution to get rid of them is to only allow void functions in a function call statement, and ban
them from expressions.

Fixes #378