Implemented non-mutating closures
This PR adds support for closures that capture, but do not re-assign to their up-values.

ir.Function now has a captured_vars (list of ir.VarDecl) and a id_of_upvalue (list of integer)field.
The captured_vars maps a local var id (which is really an upvalue) to it's decl.
The id_of_upvalue maps the local id of the upvalue to it's local id in the surrounding function.

Here is why I thought the above was necessary: In to_ir, If we tried to assign consecutive local var
ids to the arguments and the upvalues, we would have to preprocess the number of upvalues and
reserve that many id slots in loc_id_of_decl and ir.Function.vars before we start registering the
local variables. But since we're registering local variables and upvalues on the go, we may
encounter an upvalue for the first time while having assigned some local var IDs already. So
naturally we assign the next free local var id (#local_vars + 1) to the current upvalue. This means
upvalues may not have consecutive IDs after parameters in the generated C code. To prevent multiple
declarations of upvalues in the C code (once as a parameter, and once as a local variable), we make
sure captured_vars[id] is nil before emitting the declaration code for vars[id]. (Because upvalues
are not present one aftr the other following the parameter list). When capturing an upvalue (by
adding it to ccl->upvalue), we need to know the current location of an upvalue in the C stack (it's
var id), the id_of_upvalue helps with that.

In to_ir, now whenever an identifier is encountered, the order of lookup is:
    Local var
    Captured var
    Global var

In coder, some adjustments had to be made in the lua entry point code gen (to add code for capturing
the vars), gen_cmd for NewClosure (to capture the upvalues) and the pallene entry point (to add the
upvalues to the param list.)