Add compile-time assertion for luaS_newudata
To avoid integer overflow, we must ensure that the parameter to luaS_newudata is not too large. This
can happen if a record type has an unreasonably large number of GC fields, or if a program has too
many string literals.

Because USHRT_MAX is platform specific I chose to put the compile-time assertion in the C code
instead of in the Lua code.

I also decided to not bother with a nice user-friendly error message. That would be harder to do
with the assertion in the generated C code. I think this is OK because this sort of error should
only happen for pathological programs.

Closes #412.