Rename the Lua folder from "lua" to "vm"
After this PR, the custom interpreter will be found under `./vm/src/lua` instead of `./lua/src/lua`.

This change paves way the for solving issue #418. To solve that issue, we can use a Luarocks 3.0
feature where it can automatically identify all the ".lua" files in the project, provided they are
inside an appropriately named directory[1]. This would spare us from listing them one by one in the
rockspec. The problem is that one of the directories that Luarocks searches is the "lua" directory,
which we currently use for the bundled interpreter:

> To make writing rocks for projects with a common directory layout even easier, the "builtin" build
> type recursively copies all files and directories it finds in a folder called "./lua" (if it
> exists) to the same place as other Lua files. It's almost as if an entry had been added to modules
> for every file in lua, except that the files will only be copied - files ending in ".c" won't be
> compiled.

More precisely, Luarocks looks under "src", "lua", and "lib"[2], in this order. In theory, we could
take advantage of this order and put our ".lua" files under "src". However, as far as I know this
order is not officially documented. I'm also not sure if this order is the same in all versions of
Luarocks. I think it is safer if our bundled interpreter does not use these special directory names.

[1]: https://github.com/luarocks/luarocks/wiki/Creating-a-rock#building-a-module
[2]: https://github.com/luarocks/luarocks/blob/d4e82aae9f411d5d2cac10f7e043a0c3c6180eb1/src/luarocks/build/builtin.lua#L85