Revert the changes we made to the luaconf in 5.4.3
Instead of making the l_likely public in the luaconf, what we should do is mark the Pallene
generated files as being "LUA_CORE". The "define LUA_CORE" is something that already existed in
previous versions of Lua and which is present in all the core ".c" files. We probably should have
already have been using it in Pallene. It is just that this issue with the l_likely in 5.4.3 was
the first time that we noticed the problem in Pallene.