Move pallenelib to inside the Lua interpreter
This set of commits gets rid of the separate pallenelib library. This is part of the ongoing effort
to simplify the build and installation process for Pallene (see #16, #153). The pallene_core files
are now going to be part of our the custom installation of Lua.

Unfortunately, for some reason it is not working. When I compile the custom Lua, the
`pallene_symbols` are not being exported by the custom Lua executable. We'll need to figure out what
is the problem.

TODO:
- [] Find out why the symbols are not being exported
- [] Fix compilation warnings that appeared because of -Wextra
- [] Update the ".patch" file we use when updating to a new Lua version
- [] Delete the "runtime" folder, and
- [] Update the installation instructions in the README.