Linker Hack
This workaround ensures that the pallene_core. symbols are included in the compiled executable.
By default, these symbols were not being included because they are not used anywhere else in the
Lua executable. They are only intended for dynamically linked modules.

One possible way to avoid this would be to pass some special flags to the linker, such as
-Wl,--whole-archive. However, I'm worried that this might only work with a GCC toolchain on Linux.
The proposed hack is something that I hope might be more portable. However, I would certainly be
very happy if we could find an alternative less hacky solution.