New instructions for updating Pallene to the next Lua version
Previously, all we had to do was a small change in the luaconf. Now there are more things to do: add
the pallene_core files, update the makefile, add the linker hack. Because these changes broke the
"patch" file that we had, I deleted that patch file and replaced it with manual instructions.

When we update to the next release (5.4.4, or maybe 5.5?) we could consider creating a new "patch"
file. It's easier to create that patch file after we apply the changes on a fresh  version of Lua.