Move the cbenchmarks folder into the benchmark folder
The toplevel of the repository is precious; let's avoid cluttering it with things that don't need to
be there. This commit moves all the benchmarks in the cbenchmark directory into the corresponding
subdirectory of the benchmarks directory.

I believe the original motivation for separating the C benchmarks was because our benchmark runner
couldn't tell apart the two kinds of C benchmarks: C benchmarks that use the Lua API and C
benchmarks that are standalone programs. However, I don't think this is still a good reason to
maintain a separate folder.

Additionally, it makes sense to put the standalone C benchmarks together with the other ones. That
way, if one of them needs to be changed, it is easier to remember to also change the other one.