Prettier floats in generated C code
Previously, we would sometimes produce a long and ugly representation for some floating point
numbers. For example, 3.14 would become 3.1400000000000001. After this commit, we try to find the
shortest representation that is still accurate.

The new implementation might call string.format upwards of 12 times, instead of a single time.
However, it should be fast enough because this code is not performance-critical.