Continue parsing after some syntax errors
If it is obvious how to recover from a syntax error, then continue parsing.
For example, consider the following program:

    function f()
        break
    end

Previously, it would stop and complain at the start, when it failed to parse the module declaration.
Now, it continues to parse and reports several errors.

    syntax error: bar.pln:1:1: must begin with a module declaration; local <modname> = {}
    syntax error: bar.pln:2:5: break statement outside of a loop
    syntax error: bar.pln:1:1: function 'f' was not forward declared
    syntax error: bar.pln:4:1: must end by returning the module table

At the moment, the parser only continues if it is obvious how to proceed. For more complicated
syntax errors, we abort immediately because we would rather avoid producing spurious errors.
We also show an message saying that the parsing was aborted, to say that there may also be other
errors in the part of the program that we did not look at.

    syntax error: bar.pln:2:20: unexpected '%' while trying to parse a statement
    compilation aborted due to previous error

One thing that I didn't try to implement is to sort the errors so that they appear in the order
of their position in the file. Is this something that we should try to fix?