Improved error message when an "end" is missing
If an "end" token is missing, we now use an indentation-based heuristic to try to spot the actual
culprit for the error. In addition to showing the location where the error was detected, we also
show the location of the block opener that we think is to blame. For example,

    local m = {}
    local function foo()
        if true then
    end
    return m

Now shows this error:

    syntax error: foo.pln:6:1: expected 'end' before end of the file, to close the 'function' at line 2
    syntax error: foo.pln:3:5: ...possibly because this 'if' is missing an 'end' (mismatched indentation)
    compilation aborted due to previous error

We give a similar error message for mismatched parenthesis and curly braces, provided that the
open and close brace are on different lines.