Parse using "is first" instead or "not follow"
In many parts of the parser, the parser looked for the absence of a "follow" token, to decide how to
parse. This was awkward because it forced some of the functions to receive an additional argument
describing the parsing context (which kind of ruins the "context free" property of the grammar). It
also led to some strange error messages:

    local xs = {
    local y = 2

This program used to complain about "unexpected 'local' while parsing expression". Now the error says "expected '}' to close the '{'.

I do admit that there are some situations where the old error message might have been a bit clearer.
For example:

    function foo()
        1 + 1
    end

Previously, it said "unexpected number while parsing statement' and now it says "expected 'end'
before number". The new error message isn't quite right, because adding an 'end' there is not
sufficient to fix the problem...