Remove distinction between peek and peekSet
It was easy to misuse these functions by calling peek where peekSet was expected. We could have
added some extra assertions in peek and peekSet, but if we're doing to check the type of the
argument then we might as well make the function polymorphic...

As a bonus, this also makes it possible to do a doublepeekSet.