Recompile benchmarks when running test suite
Because the test suite did not recompile the benchmarks, sometimes the tests would pass on the local
machine but not pass on the CI server. The immediate workaround is to force recompilation when
running the test suite. In the future it would be nice if we replaced the benchlib by something less
crazy.