Rename pallene_log to pallene_math_log
I felt that pallene_log sounded like it was a logging function, instead of a math function. Let's
make the name a bit more explicit...

Also, add a comment saying that the code is based on the one from lmathlib.c, so we know where to
check the next time when we port Pallene to a future Lua version.