Conform to latest Luacheck
Our CI started failing because some recent version of Luacheck (don't know which one) changed how it
deals with unused variables. This is what I noticed:

1. By default, Luacheck now considers that variables starting with '_' can be unused. We no longer
need to handle this special case in our luacheckrc.
2. There is a new warning category (214) that complains about using a '_' variable. There were two
places in our code that did this. The first was a _recursive parameter, which I renamed without the
underscore. The second was _ENV, which I told luacheck to ignore in the configure file.