Make Pallene installable with Luarocks
Exciting news! This commit cleans up the way we install Pallene and
resolves several long-standing issues:

* Closes #16
* Closes #18
* Closes #153

Now Pallene can be installed like a regular Luarocks package, using
`luarocks make`. To achieve this I did two things: get rid of #include
directives in the generated C code, and get rid of the custom Lua
interpreter.

Include directives were annoying because they meant you had to pass the
right "-I" flags when you compiled the generated C code. Also, we never
figured out a good place to install the headers in the first place,
which was one of the reasons for #16... My approach to solve this
problem was to get rid of the #include directives completely, fully
expanding the header files. This required introducing a code generation
step to build the Pallene library.

The main thing that motivated me to get rid of the custom Lua is because
a custom Lua also implies a custom Luarocks, and I wanted to make things
as easy as possible to install via Luarocks. The catch, however, is that
we have a requirement saying that Lua needs to be exactly 5.4.4. This is
OK if you're using Fedora, which always has the latest Lua, but on most
other distros you're might need to install Lua from source...

To make things less magical and more compatible with Luarocks, we
implemented the Pallene runtime library as a struct of function
pointers, stored inside an userdata and installed in an extention module
that can be required. This design, without "extern" functions,
completely bypasses the C linker. We no longer need to worry about that
passing the right linker flags to the compiler, and we can get rid of
the "linker hack" that we had added to lapi.c. All in all, the C files
we generate are now a bit more verbose, but I think it is for the
better.