Include entire Pallene lib into the generated C
This makes it easier to compiler the generated C code because it no
longer needs to #include the pallene_core.h, passing the right -I flags.
We also don't have to compile & install the Pallene standard library,
meaning that the Pallene compiler is now a pure Lua project.

The catch here is that we now statically link the Pallene standard
library, which adds up to 10 KB of extra size to every binary produced
with Pallene. I think that is an acceptable tradeoff for now, to
keep things simple. If this turns out to be a problem later then we can
reintroduce the pallenelib.so.