Move the dev-minded CFLAGS into the run-tests
This removes debugging (-g) and warning (-W) flags in a default
invocation of pallenec. Again, the main reason is to pave the road for
the Luarocks compilation backend, which requires that any custom flags
be passed via the CFLAGS environment variable. However, I think it is
still a good idea despite that. Removing -g leads to smaller binaries
and faster compilation times. As for the warning-related flags, they are
not relevant to Pallene end-users, they are only relevant to us who are
writing the compiler.