git-sync: Use "git branch -d" to delete branches (#535)
This patch fixes a bug with our git-sync script, which caused it to
fail to properly dispose of the deleted branches. **If you were using
this script, you must clone fresh copy of the repository from Github,**
in order to get rid of the garbage branch data that has accumulated.

The problem was that we used "git update-ref -d" to delete the local
branches. That deleted the local branch but did not delete the tracking
information between the local branch and the local branch (that one
that you set using "git push -u"). The result is that over time the
.gitconfig accumulates this tracking info that never gets deleted. I
noticed it for the first time when I tried to create a new branch with a
name that had previously been used in a since-deleted branch. I got the
following error message:

    Your branch is based on 'origin/lua-core', but the upstream is gone.